requirements.txt
# Core
pytz==2021.3
Django==4.0.1
django-configurations==2.3.1
gunicorn==20.1.0
newrelic==7.2.4.171
django-cors-headers

#Libs
google-api-python-client==2.36.0
facebook-sdk-python


# For the persistence stores
psycopg2-binary==2.9.3
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.2.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.13.1
djangorestframework-simplejwt
Markdown==3.3.6
django-filter==21.1
django-oauth-toolkit==1.7.0
djangorestframework-oauth==1.1.0
drf_social_oauth2

# Developer Tools
ipdb==0.13.9
ipython==8.0.0
mkdocs==1.2.3
flake8==4.0.1
drf-yasg
tox


# Testing
mock==4.0.3
factory-boy==3.2.1
django-nose==1.4.7
nose2
coverage==6.2
model_bakery==1.4.0
selenium==4.1.0
black

# Static and Media Storage
django-storages==1.12.3
boto3==1.20.37

# AP ID
# 707890890378344